库存下单

+++ app/web/src/main/java/com/zeh/wms/web/controller/api/model/OrderCalculateModel.java
@@ -17,7 +17,7 @@ public class OrderCalculateModel implements Serializable {
 
     private Long              commodityId;
 
-    private Double            commodityQuantity;
+    private Integer           commodityQuantity;
 
     private String            expressType;
 }
